Create Radius
You are to write a program that prompts the user to enter a radius of a circle/sphere. Using the radius that is entered, the program must calculate and display the circumference and area of a circle and the volume of a sphere with that radius and then print the results to the screen.